+++ runtime/src/main/java/ru/hollowhorizon/hollowengine/common/scripting/DefaultScriptDefinitions.kt
@@ -62,6 +62,7 @@ object DefaultScriptDefinitions {
                     "ru.hollowhorizon.hollowengine.common.ui.hud.*",
                     "ru.hollowhorizon.hollowengine.common.ui.net.*",
                     "ru.hollowhorizon.hollowengine.client.ui.script.observe",
+                    "ru.hollowhorizon.hollowengine.client.ui.script.LocalReplacedScreen",
                     "ru.hollowhorizon.hollowengine.common.scripting.annotations.*",
                     "net.minecraft.nbt.CompoundTag",
                     ResourceLocation::class.qualifiedName!!,